vending-machine/requirements.txt
python-dotenv
asyncio
python-socketio[asyncio_client]
pyjwt
RPi.GPIO
